refactor(chips): represent internal amounts as int64_t table chips
Introduce a single normalization factor (1 CHIP = 100 table chips) so the
backend stores all in-game amounts as int64_t instead of double CHIPS.
Conversion to/from on-chain CHIPS happens only at the boundary:

  - INI parse (config.c) and cashier defaults (cashier.c) convert CHIPS
    doubles to table-chip int64_t at load time
  - vdxf.c converts payin CHIPS read from on-chain transactions into
    table chips before storing in the payin_amounts CMM, and converts
    table_min_stake back to CHIPS for the on-chain sendcurrency
  - blinder.c converts settle_amounts back to CHIPS for payout txs
  - host.c legacy multisig path converts at the wire boundary

GUI receives integer table chips ({"pot":4,"toCall":1,"player_funds":[50,50]}),
which avoids the Stack.tsx fractional-amount crash without touching GUI code.
The Verus ID CMM payload now serializes integer-valued doubles
(e.g. {"amount": 2}) for action amounts, betting_state and settlement.

Helpers `chips_to_table_chips()` / `table_chips_to_chips()` live in common.h.
struct privatebet_vars amount fields (pot, funds, betamount, win_funds, etc.)
are int64_t. Globals renamed: SB_in_chips -> SB_in_table_chips, BB_in_chips ->
BB_in_table_chips, table_min_stake -> table_min_stake_in_table_chips, etc.
Logging shows both formats, e.g. "1 table chips (0.0100 CHIPS)".

Verified end-to-end with CLI auto: dealer reset, 2-player join, blinds,
hole cards, flop/turn/river, and showdown all flow with clean integer
chip values; settlement payload pot=4, settle_amounts=[50,50] written
correctly. (GAP D settlement deadlock and GAP G/H GUI->VDXF write are
pre-existing and unaffected by this change.)

+++ poker/include/bet.h
@@ -112,19 +112,23 @@ struct privatebet_vars {
 	cJSON *actions[CARDS_MAXROUNDS][CARDS_MAXPLAYERS + 1];
 	int32_t mypermi[CARDS_MAXCARDS], permi[CARDS_MAXCARDS], turni, round, validperms, roundready, lastround,
 		numconsensus;
-	double small_blind, big_blind;  // Actual CHIPS values (e.g., 0.01, 0.02)
-	double betamount[CARDS_MAXPLAYERS][CARDS_MAXROUNDS];
+	/* All chip-denominated fields are integer "table chips" (1 CHIP = 100
+	 * table chips). Conversion happens only at the on-chain boundaries -
+	 * see chips_to_table_chips() / table_chips_to_chips() in common.h.
+	 * Default blinds: small_blind=1, big_blind=2 (i.e. 0.01 / 0.02 CHIPS). */
+	int64_t small_blind, big_blind;
+	int64_t betamount[CARDS_MAXPLAYERS][CARDS_MAXROUNDS];
 	int32_t bet_actions[CARDS_MAXPLAYERS][CARDS_MAXROUNDS];
 	int32_t dealer, last_turn;
-	double last_raise;
+	int64_t last_raise;
 	int64_t turn_start_time;    // Unix timestamp when current turn started
 	int32_t turn_start_block;   // Block height when current turn started
-	double pot;
-	double player_funds;
-	double funds[CARDS_MAXPLAYERS];       // Available funds in CHIPS
-	double funds_spent[CARDS_MAXPLAYERS]; // Total spent this game
-	double win_funds[CARDS_MAXPLAYERS];   // Winnings from pot
-	double ini_funds[CARDS_MAXPLAYERS];   // Initial funds (payin amount)
+	int64_t pot;
+	int64_t player_funds;
+	int64_t funds[CARDS_MAXPLAYERS];       // Available funds (table chips)
+	int64_t funds_spent[CARDS_MAXPLAYERS]; // Total spent this game
+	int64_t win_funds[CARDS_MAXPLAYERS];   // Winnings from pot
+	int64_t ini_funds[CARDS_MAXPLAYERS];   // Initial funds (payin amount)
 	int32_t winners[CARDS_MAXPLAYERS];
 	char player_chips_addrs[CARDS_MAXPLAYERS][64];
 	int32_t req_id_to_player_id_mapping[CARDS_MAXPLAYERS];
